調整項目工程結構
